--- agents/agents-core/src/jvmCommonMain/kotlin/ai/koog/agents/planner/goap/ActionBuilder.kt
@@ -1,62 +1,0 @@
-@file:Suppress(
-    "MissingKDocForPublicAPI",
-    "EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING",
-    "ACTUAL_ANNOTATIONS_NOT_MATCH_EXPECT"
-)
-
-package ai.koog.agents.planner.goap
-
-import ai.koog.agents.annotations.JavaAPI
-import ai.koog.agents.core.agent.context.AIAgentPlannerContext
-import ai.koog.agents.core.annotation.InternalAgentsApi
-import ai.koog.utils.annotations.InternalKoogUtils
-import ai.koog.utils.concurrency.withContextReentrant
-
-@OptIn(InternalKoogUtils::class, InternalAgentsApi::class)
-public actual class ActionBuilder<State> : ActionBuilderApi<State> {
-    private val delegate = ActionBuilderImpl<State>()
-    public actual override fun name(name: String): ActionBuilder<State> = apply { delegate.name(name) }
-    public actual override fun description(description: String?): ActionBuilder<State> =
-        apply { delegate.description(description) }
-
-    public actual override fun precondition(precondition: Condition<State>): ActionBuilder<State> =
-        apply { delegate.precondition(precondition) }
-
-    public actual override fun belief(belief: Belief<State>): ActionBuilder<State> = apply { delegate.belief(belief) }
-    public actual override fun cost(cost: Cost<State>): ActionBuilder<State> = apply { delegate.cost(cost) }
-    public actual override fun execute(execute: Execute<State>): ActionBuilder<State> =
-        apply { delegate.execute(execute) }
-
-    /**
-     * Sets the synchronous execute function for the action.
-     */
-    @JavaAPI
-    @Deprecated("Use execute(ExecuteSync) instead.", ReplaceWith("execute(execute)"))
-    public fun executeSync(execute: ExecuteSync<State>): ActionBuilder<State> =
-        execute { context, state ->
-            withContextReentrant(context.config.strategyDispatcher) {
-                execute.execute(context, state)
-            }
-        }
-
-    /**
-     * Sets the synchronous execute function for the action.
-     */
-    @JavaAPI
-    @JvmName("execute")
-    public fun javaApiExecuteSynchronously(execute: ExecuteSync<State>): ActionBuilder<State> =
-        execute { context, state ->
-            withContextReentrant(context.config.strategyDispatcher) {
-                execute.execute(context, state)
-            }
-        }
-
-    public actual override fun build(): Action<State> = delegate.build()
-
-    /**
-     * Synchronous GOAP action execution.
-     */
-    public fun interface ExecuteSync<State> {
-        public fun execute(context: AIAgentPlannerContext, state: State): State
-    }
-}
