+++ agents/agents-planner/build.gradle.kts
@@ -1,0 +1,41 @@
+import ai.koog.gradle.publish.maven.Publishing.publishToMaven
+
+group = rootProject.group
+version = rootProject.version
+
+plugins {
+    id("ai.kotlin.multiplatform")
+    alias(libs.plugins.kotlin.serialization)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(project(":agents:agents-core"))
+                implementation(libs.kotlinx.serialization.json)
+                implementation(libs.oshai.kotlin.logging)
+            }
+        }
+
+        commonTest {
+            dependencies {
+                implementation(project(":agents:agents-test"))
+                implementation(project(":test-utils"))
+                implementation(libs.kotest.assertions.json)
+            }
+        }
+
+        jvmTest {
+            dependencies {
+                implementation(kotlin("test-junit5"))
+                implementation(project(":prompt:prompt-executor:prompt-executor-clients:prompt-executor-ollama-client"))
+                implementation(project(":serialization:serialization-jackson"))
+            }
+        }
+    }
+
+    explicitApi()
+}
+
+publishToMaven()
