+++ agents/agents-planner/src/commonMain/kotlin/ai/koog/agents/planner/goap/ActionBuilderImpl.kt
