+++ agents/agents-planner/src/jvmTest/java/ai/koog/agents/planner/goap/GoalBuilderJavaTest.java
