+++ integration-tests/build.gradle.kts
@@ -25,6 +25,7 @@ kotlin {
         jvmTest {
             dependencies {
                 implementation(project(":agents:agents-ext"))
+                implementation(project(":agents:agents-planner"))
                 implementation(project(":agents:agents-features:agents-features-event-handler"))
                 implementation(project(":agents:agents-features:agents-features-trace"))
                 implementation(project(":agents:agents-features:agents-features-snapshot"))
