+++ koog-agents/build.gradle.kts
@@ -66,6 +66,7 @@ val included = setOf(
     ":agents:agents-features:agents-features-snapshot",
     ":agents:agents-mcp",
     ":agents:agents-mcp-metadata",
+    ":agents:agents-planner",
     ":agents:agents-tools",
     ":agents:agents-utils",
     ":embeddings:embeddings-base",
