+++ settings.gradle.kts
@@ -107,3 +107,5 @@ include(":docs")
 
 include(":test-utils")
 include(":utils")
+
+include(":agents:agents-planner")
